work on deploy command and add deploy Cloud Function functionality

+++ src/gcp/destroyBuckets.js
@@ -1,5 +1,6 @@
 const {Storage} = require('@google-cloud/storage');
 const storage = new Storage();
+const { bucketExists } = require('./validateIntegrationName');
 
 const destroyBuckets = async (integrationName) => {
   await Promise.all([destroyBucket(integrationName),
@@ -8,7 +9,13 @@ const destroyBuckets = async (integrationName) => {
 
 const destroyBucket = async (integrationName) => {
   const bucket = storage.bucket(integrationName);
-  await bucket.delete();
+  
+  if (await bucketExists(integrationName)) {
+    await bucket.delete();
+    console.log(`Bucket "${integrationName}" is destroyed.`);
+  } else {
+    console.log(`Bucket "${integrationName}" doesn't exist!`);
+  }
 }
 
 module.exports = {
